+++ src/core/containers/project/ProjectListing.js
@@ -21,29 +21,28 @@ import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import { Button, Icon, Table } from 'semantic-ui-react'
 import {
-    deleteProject, fetchProjects, projectDeleteError, projectEditErrorInListing,
-    updateProjectNameInListing
+    clearProjectActionError, createProject, deleteProject, fetchProjects,
+    toggleShowProjectForm
 } from '../../actions/project/ProjectListing'
-import CreateProjectForm from './CreateProjectForm'
 import { IconButton } from '../../components/Button'
 import { ErrorMessage } from '../../components/Message';
 import DeleteResourceModal from '../../components/modals/DeleteResourceModal'
-import EditResourceNameModal from '../../components/modals/EditResourceNameModal'
+import CreateProjectForm from '../../components/project/CreateProjectForm'
 import ContentSpinner from '../../components/ContentSpinner'
-import { isNotEmptyString, pageUrl } from '../../util/App.js';
+import { pageUrl } from '../../util/App.js';
 
 import '../../../css/ResourceListing.css'
 
 
 class ProjectListing extends Component {
     static propTypes = {
+        actionError: PropTypes.object,
         envs: PropTypes.array,
-        deleteError: PropTypes.string,
-        editError: PropTypes.string,
         fetchError: PropTypes.string,
         isFetching: PropTypes.bool.isRequired,
         projects: PropTypes.array.isRequired,
-        links: PropTypes.object
+        links: PropTypes.object,
+        showForm: PropTypes.bool.isRequired
     }
     constructor(props) {
         super(props);
@@ -51,25 +50,18 @@ class ProjectListing extends Component {
         // or edit. If either value is non-null the respective modal to confirm
         // delete or edit the project is shown. At no point should both values
         // be non-null.
-        this.state = {deleteProject: null, editProject: null}
+        this.state = {deleteProject: null}
     }
     /**
      * Load the project listing when the component mounts.
      */
     componentDidMount = () => (this.refresh());
     /**
-     * Clear delete project error message
+     * Clear create or delete project error message.
      */
-    clearDeleteProjectError = () => {
+    clearActionError = () => {
         const { dispatch } = this.props
-        dispatch(projectDeleteError(null))
-    }
-    /**
-     * Clear delete project error message
-     */
-    clearEditProjectError = () => {
-        const { dispatch } = this.props
-        dispatch(projectEditErrorInListing(null))
+        dispatch(clearProjectActionError());
     }
     /**
      * Delete the project with the given Url.Delete
@@ -77,13 +69,13 @@ class ProjectListing extends Component {
     confirmDeleteProject = (project) => {
         const { dispatch } = this.props
         dispatch(deleteProject(project))
-        this.hideModals()
+        this.hideModal()
     }
     /**
      * Hide all modals by setting the respective state variables to null..
      */
-    hideModals = () => {
-        this.setState({deleteProject: null, editProject: null})
+    hideModal = () => {
+        this.setState({deleteProject: null})
     }
     /**
      * Re-load project lsiting (either when component mounts or when user
@@ -98,7 +90,14 @@ class ProjectListing extends Component {
      * 'Delete Project' dialog is being displayed.
      */
     render() {
-        const { isFetching, deleteError, editError, envs, fetchError, projects } = this.props;
+        const {
+            isFetching,
+            actionError,
+            envs,
+            fetchError,
+            projects,
+            showForm
+        } = this.props;
         let content = null;
         if (isFetching) {
             content = <ContentSpinner text='Loading Projects ...'/>;
@@ -112,12 +111,10 @@ class ProjectListing extends Component {
                     <Table.Row>
                         <Table.HeaderCell className="resource">Name</Table.HeaderCell>
                         <Table.HeaderCell className="resource">Project type</Table.HeaderCell>
-                        <Table.HeaderCell className="resource">Created at</Table.HeaderCell>
                         <Table.HeaderCell className="resource">Last modified</Table.HeaderCell>
                         <Table.HeaderCell className="resource"></Table.HeaderCell>
                     </Table.Row>
             );
-            projects.sort(function(p1, p2) {return p1.name.localeCompare(p2.name)});
             let rows = [];
             for (let i = 0; i < projects.length; i++) {
                 const pj = projects[i];
@@ -129,15 +126,8 @@ class ProjectListing extends Component {
                     <Table.Cell className={'resource-text'}>
                         {envs.find(e => (e.id === pj.envId)).name}
                     </Table.Cell>
-                    <Table.Cell className={'resource-text'}>{pj.createdAt}</Table.Cell>
                     <Table.Cell className={'resource-text'}>{pj.lastModifiedAt}</Table.Cell>
                     <Table.Cell className={'resource-buttons'}>
-                        <span className='button-wrapper'>
-                            <IconButton name="edit" onClick={(event) => {
-                                event.preventDefault();
-                                this.showEditProjectModal(pj);
-                            }}/>
-                        </span>
                         <span className='button-wrapper'>
                             <IconButton name="trash" onClick={(event) => {
                                 event.preventDefault();
@@ -147,11 +137,10 @@ class ProjectListing extends Component {
                     </Table.Cell>
                 </Table.Row>);
             }
-            // Check if deleteProject or editProject are set. In that case
-            // display a modal dialog for the user to either confirm (or cancel)
-            // project deletion or enter the new Pproject name. It is assumed
-            // that at most one of the two values is non-null.
-            const { deleteProject, editProject } = this.state
+            // Check if deleteProject is set. In that case display a modal
+            // dialog for the user to either confirm (or cancel) project
+            // deletion.
+            const { deleteProject } = this.state
             let modal = null;
             if (deleteProject) {
                 modal = (
@@ -160,19 +149,8 @@ class ProjectListing extends Component {
                         prompt={'Do you really want to delete ' + deleteProject.name + '?'}
                         title='Delete Project'
                         value={deleteProject}
-                        onCancel={this.hideModals}
+                        onCancel={this.hideModal}
                         onSubmit={this.confirmDeleteProject}
-                    />
-                );
-            } else if (editProject) {
-                modal = (
-                    <EditResourceNameModal
-                        isValid={isNotEmptyString}
-                        open={true}
-                        title='Enter a new project name ...'
-                        value={editProject.name}
-                        onCancel={this.hideModals}
-                        onSubmit={this.submitProjectNameUpdate}
                     />
                 );
             }
@@ -180,27 +158,33 @@ class ProjectListing extends Component {
             // project. The message is shown at the bottom of the project
             // listing
             let projectActionErrorMessage = null
-            if (deleteError) {
+            if (actionError) {
                 projectActionErrorMessage = (<ErrorMessage
-                    title="Error while deleting project"
-                    message={deleteError}
-                    onDismiss={this.clearDeleteProjectError}
+                    title={actionError.title}
+                    message={actionError.message}
+                    onDismiss={this.clearActionError}
                 />)
-            } else if (editError) {
-                projectActionErrorMessage = (<ErrorMessage
-                    title="Error while updating project name"
-                    message={editError}
-                    onDismiss={this.clearEditProjectError}
-                />)
+            }
+            let createProjectForm = null;
+            if (showForm) {
+                createProjectForm = (
+                    <CreateProjectForm
+                        envs={envs}
+                        onClose={this.toggleCreateProjectForm}
+                        onSubmit={this.submitNewProject}
+                    />
+                );
             }
             content = (
                 <div>
+                    { projectActionErrorMessage }
+                    { createProjectForm }
                     <Table singleLine>
                         <Table.Header>{tabHead}</Table.Header>
                         <Table.Body>{rows}</Table.Body>
                         <Table.Footer fullWidth>
                             <Table.Row>
-                                <Table.HeaderCell colSpan='5'>
+                                <Table.HeaderCell colSpan='4'>
                                     <Button
                                         floated='right'
                                         icon
@@ -210,13 +194,22 @@ class ProjectListing extends Component {
                                     >
                                       <Icon name='refresh' /> Refresh
                                     </Button>
+                                    <Button
+                                        floated='right'
+                                        icon
+                                        labelPosition='left'
+                                        size='tiny'
+                                        positive
+                                        disabled={showForm}
+                                        onClick={this.toggleCreateProjectForm}
+                                    >
+                                      <Icon name='plus' /> New Project ...
+                                    </Button>
                                 </Table.HeaderCell>
                             </Table.Row>
                             </Table.Footer>
                     </Table>
                     { modal }
-                    { projectActionErrorMessage }
-                    <CreateProjectForm />
                 </div>
             );
         }
@@ -226,35 +219,40 @@ class ProjectListing extends Component {
      * Show modal to confirm project deletion.
      */
     showDeleteProjectModal = (project) => {
-        this.setState({deleteProject: project, editProject: null})
+        this.setState({deleteProject: project})
     }
     /**
-     * Show modal to edit project name.
+     * Submit a create new project request. If the name is empty it is set to
+     * 'undefined' by default.
      */
-    showEditProjectModal = (project) => {
-        this.setState({deleteProject: null, editProject: project})
+    submitNewProject = (name, env) => {
+        const { dispatch, links } = this.props;
+        this.toggleCreateProjectForm();
+        let projectName = name.trim();
+        if (projectName === '') {
+            projectName = 'New Project';
+        }
+        dispatch(createProject(links.create, env, projectName));
     }
     /**
-    * Submit the update file name request.
-    */
-   submitProjectNameUpdate = (name) => {
-       const {dispatch } = this.props
-       const { editProject } = this.state
-       dispatch(updateProjectNameInListing(editProject, name))
-       this.hideModals()
-   }
+     * Toggle visibility of the create project form.
+     */
+    toggleCreateProjectForm = () => {
+        const { dispatch } = this.props;
+        dispatch(toggleShowProjectForm())
+    }
 }
 
 const mapStateToProps = state => {
 
     return {
+        actionError: state.projectListing.actionError,
         envs: state.projectListing.envs,
         fetchError: state.projectListing.fetchError,
-        deleteError: state.projectListing.deleteError,
-        editError: state.projectListing.editError,
         isFetching: state.projectListing.isFetching,
         projects: state.projectListing.projects,
-        links: state.projectListing.links
+        links: state.projectListing.links,
+        showForm: state.projectListing.showForm
     }
 }