+++ src/core/reducers/main/MainPage.js
@@ -16,32 +16,28 @@
  * limitations under the License.
  */
 
-import { SET_ACTIVE_ITEM } from '../../actions/main/MainPage'
 import { RECEIVE_SERVICE } from '../../actions/main/Service'
-import { MENU_ITEM_HOME } from '../../containers/main/MainPage'
 
-/**
-* Reducer for actions that manipulate the global state of the App home page.
-*/
 
 /**
  * STATE:
  *
- * activeItem: name of the selected menu item
+ * homePageContent: Welcome message {title: ..., description: ...} provided by
+ *    the API.
+ * projects: Project listing returned by the API.
  */
 
 const INITIAL_STATE = {
-    activeItem: MENU_ITEM_HOME,
-    homePageContent: null
+    homePageContent: null,
+    projects: []
 }
 
+
+/**
+* Reducer for actions that manipulate the global state of the App home page.
+*/
 export const mainPage = (state = INITIAL_STATE, action) => {
     switch (action.type) {
-        case SET_ACTIVE_ITEM:
-            return {
-                ...state,
-                activeItem: action.item
-            }
         case RECEIVE_SERVICE:
             return {
                 ...state,